feat(render): update the whole rendering process and enable to use brightness in all pixels
Many thanks to Julien Dorra <juliendorra@juliendorra.com> and Jonathan Perret <j-github@jonathanperret.net> for their work on the rendering!

+++ include/mode/snake.h
@@ -7,17 +7,17 @@
 class Snake
 {
 private:
-  static const unsigned int LED_TYPE_OFF = 0;
-  static const unsigned int LED_TYPE_ON = 1;
-  static const unsigned int GAME_STATE_RUNNING = 1;
-  static const unsigned int GAME_STATE_END = 2;
+  static const uint8_t LED_TYPE_OFF = 0;
+  static const uint8_t LED_TYPE_ON = 1;
+  static const uint8_t GAME_STATE_RUNNING = 1;
+  static const uint8_t GAME_STATE_END = 2;
 
   byte gameState;
   byte lastDirection = 0; // 0=unset 1=up 2=right 3=down 4 =left
 
   std::vector<uint> position = {240,241,242};
 
-  unsigned int dot;
+  uint8_t dot;
 
   void initGame();
   void newDot();
